SupportMediaDemos: Fix crash of VideoPlayerActivity in multi-window
VideoViewImplBase.attachMediaControlView() assumes that mMediaControlView is not null.
However, if the attribute widget:enableControlView of VideoView is set to false,
mMediaControlView is not created and is null.

Usually onAttachedToWindow() is called after an activity is created(after
onCreate()), but it is called during setContentView() when an app exits multi-window mode.

In this CL, we checks if mMediaControlView is null to prevent the crash.

Bug: 122285080
Test: ran support-media-demos manually
Change-Id: I44bd81db075c0f7b9bde095871b1f5857b3cc043