Introduce Density class
We already have too many usages of Context in our classes just to convert dimensions to pixels. It is time to create an abstraction for dimensions to not depend on Android in our Crane modules. Density is just a data class with one float value. We will be able to make it an inline class when it stops crashing for our experimental Kotlin version.

Bug: 123282488
Test: run our tests and manually
Change-Id: I7f5238639240a3e90a8e758f84a80cf3924492ba