Improved error messages for using wrong FragmentManager
Fragments should be associated with one FragmentManager at a time. We
should provide better error messages to developers when this construct
is not followed. Unless the interaction is add/attach/replace, we should
throw error messages if a Fragment is trying to interact with a
FragmentManager other than the one it is attached to. For Fragments with
no associated FragmentManager all interactions are allowed.

Test: tests have been added to BackStackStateTest.kt
BUG: 73830568
Change-Id: Ibf96623e0512ecc30aca8c5273bc2a5437b3b4de