Fix SampleDyanmicGroupMRP
Before aosp/831900, a dynamic group is created when the first route is selected
but that CL changed the behavior such that a dynamic group is created
just before a member route is added or removed.(during the second
action).

Since the variable mTvSelectedCount was valid after a dynamic group is
created, there was a chance that two or more TVs are selected.

Also, mMemberRouteIds was valid after a dynamic group is created so the
users couldn't remove a member from a static group.

This CL initializes the variables so it disables another TV to be
selected and enables a member from a static gorup can be removed.

Bug: 124482961
Test: ran v7 support demos manually
Change-Id: I5babffa2a8af2b6f92d2b581ab800aad5583957e