Improve GhostView logic on APIs < 21
Starting from API 21 we use platform's GhostView which has this features comparing to the current ported version:
1) If you add multiple views as a ghosts for the same parent they would be added in the order in which they would be drown in the original hierarchy
2) They will also be drawn on top of other overlay views(not ghosts).
We reimplement this logic in the ported version and start using this on all APIs expect Pie. We can't use it on Pie as we can't enable a shadow support for a canvas on this API version.
Plus I fixed wrong insertion ordering for ViewOverlayApi14 and wrong matrix calculation for transformMatrixToLocal on APIs < 21.

Bug: 125777978
Bug: 117521193
Bug: 123768642
Test: added new tests for GhostView which verifies the correct drawing order of the views
Change-Id: Ie8a0c952cd15b2c3d29a816c02e49a8adff85a96