Make ViewModelProvider.KeyedFactory package private
Separate interface for KeyedFactory doesn't compose: it makes to double
surface API, we should have both methods one that receives Factory and
one that receives KeyedFactory. In addition in kotlin extensions we have
a fun that receives `block: () -> Factory`, in this case we can't even
duplicate, because `block: () -> KeyedFactory` has same JVM signature.

So we hide it from public API.

Test: ViewModelProviderTest
Change-Id: Ifd9f21c7216b212e316c977538c2b9f71df6aa41