Add withDensity effect and a dimension conversion receiver scope.
I moved all the dimension conversions into DensityReceiver interface as functions with a default implementation.
Now we can get rid of most of the direct usages of the Density object.
We can also easily add this conversion functions into any Receiver scope with just implementing this interface.
In that way I enabled conversions for MeasureBox and ComplexMeasureBox receiver scopes and also started using DensityReceiver as a receiver scope for Draw composable which simplifies a lot of code in the material module.

Now when you are
1) inside MeasureBox or Draw you just have this functions.
2) somewhere outside but within a composition you can use an effect +withDensity { // your conversion }. Please take a look on an example in Container class.
3) somewhere outside of the composition but have a density object you can use withDensity(density) { // your conversion } syntax.

Bug: 127630050
Test: sample app and run tests
Change-Id: I657906373fe29d257331213f54bfc06b64bfb6c8