Allow extra layout space on both sides
Deprecate the old method getExtraLayoutSpace(state) in favor of a new
mechanism that allows to have custom extra layout space on both sides.

The new method is calculateExtraLayoutSpace(state, int[]), which
calculates the extra layout space on the start and end side of the
viewport ([0] and [1], respectively). Backwards compatibility is
provided by calling getExtraLayoutSpace from the default implementation.
This will behave exactly as before if the new method is not overridden.

Bug: 114217017
Test: ./gradlew recyclerview:connectedCheck
Change-Id: I4a2fc45978ea3304cb9ee581e83739a44ce37a5a