Make setDefaultNightMode immediately apply
Currently setDefaultNightMode is just a simple setter,
meaning that it only takes effect in any components
which are created after the call. This is different
to setLocalNightMode() which applies the new night
mode immediately.

This CL syncs up the behavior between the two, so that
any AppCompatDelegates in the 'created' state have their
DayNight mode re-applied.

Also tidied up the method docs, since setDefaultNightMode
is now the preferred way to control DayNight.

BUG: 129365513
Test: updated current tests to parametize over both local/default
Test: ./gradlew appcompat:connectedCheck

Change-Id: I426c81567a3148afae5b9e4d44b6eeb7f4870ef1