AccessibilityEvents from the RecyclerView shouldn't populate text from all of it's children.
NOTE: Uploaded with --no-verify to workaround bug, be extra vigilent here.

The nature of a recyclerView means, that this text is probably gonna be too long and not helpful.
FOr this particular bug, the ReyclerView took focus, which triggered a TYPE_VIEW_FOCUSED AccessibilityEvent.
This boils down to a call to dispatchPopulateAccessibilityEvent, where the ViewGroup implementation calls dispatchPopulateAccessibilityEvent on all of the children.
The goal of this API is to populate the text field in AccessibilityEvent, so the result is that this AccessibilityEvent speaks the text of all contained (but visible) views.
It creates a long non-sensical string that isn't interupted in  Talkback (which is a seperate issue for the Talkback team to address).
We have an open bug to deprecate dispatchPopulateAccessibilityEvent, and make the services use something else, but until then this fix will have to do.

Test: RecyclerViewAccessibilityTest (I didn't build the settings app with this fix, but did rebuild frameworks with a similar fix to see if my theory is correct).
Change-Id: I5a47fef0b16c8b09463c2143855c721923348e3e
Fix: 128386269