Update Ripples to use Swan transitions
Plus made few fixes in TransitionAnimations to better handle some asynchronous events like the new state can be requested before we will have the first frame callback. It also helped to fix the Animation Demo/State Based Ripple as it became broken because of the same issue after update from MeasureBox to Layout in this sample.

Bug: 114021677
Test: tested in our demo app
Change-Id: I7e3257b0abade96e835a7309319a631b2185a374