Allow navigating in app via deep link URIs
Adds API to call navigate() with a deeplink. It also allows for navigation
to destinations within other graphs (as long as the graphs are included
together by higher level modules). A new destination is added to the back
stack and popping it returns to the previous destination, even if it is
in a different graph.

Test: new tests and ran nav controller tests
	./gradlew checkApi
BUG: 110412864
Change-Id: Ie88f3ab8359c6a9fde28f0fc29c1917cc23f41f7