Deprecate Cancellable
Cancellable is no longer being used in
public APIs and suffers from a lack of
composable infrastructure that would be
necessary for it to be a general purpose
solution that we should recommend to
developers.

Deprecate the public Cancellable API and
move internal usage of it in androidx.activity
to a package private interface so as to avoid
any cross group id dependency.

Test: ./gradlew checkApi
BUG: 129946737
Change-Id: I8c6c40fc9dfc8562584bd398c4ba33fe6cc18f3b