Make Room an isolating incremental annotation processor.
This CL introduces an annotation processor option called
    `room.incremental`
which the user can set to try out incremental annotation processing with
Room. It is currently off by default.

The user can pass the option through the Android Gradle plugin's DSL,
for example:

android {
    defaultConfig
        javaCompileOptions {
	    annotationProcessorOptions {
                argument 'room.incremental', 'true'
            }
	}
    }
}

Caveat: There is a JDK bug
https://bugs.openjdk.java.net/browse/JDK-8007720 that prevents Room from
being incremental (details are discussed in the code comment).
Therefore, if the JDK version being used does not already have this fix,
Room cannot be incremental and we will print out a warning to ask the
user to use a JDK version having the fix.

Note that Room currently does not track the annotation processor option
`room.schemaLocation` properly as an @OutputDirectory, and this could
impact build correctness. However, this issue impacts both
non-incremental and incremental builds, and it already existed before,
so we will fix it separately (tracked at bug 132245929).

Additionally, Room is currently not using the
javax.annotation.processing.Filer API to let Gradle know the originating
element of a generated schema file, so the schema files may not get
deleted properly when certain source files are deleted (tracked at bug
134472065).

Bug: 112110217
Test: New RoomIncrementalAnnotationProcessingTest
Change-Id: I0076e596b1c05bcc51afea21a4f6a9a90fb6c7be