Disable double activity animation when opening app with deeplink
When using handleDeepLink(Intent) to open an app, Navigation enforces
that Intents that set the NEW_TASK flag also set the CLEAR_TASK flag.
This ensures a consistent activity back stack, and means that the
Activity should be created twice when the calling app only uses
NEW_TASK. Both Activity creations are currently being animated, causing
what appears as a flash before the app starts.

This change ensures that the second Activity creation animation is
disabled.

Test: Tested in app, all tests pass.
BUG: 130362979
Change-Id: Ie765aea34176508b7c1b4b897ca25e0d11b2d819