Create database dir when copying database.
On APIs below 24, context.getDatabaseFile() does not automatically
create the app's internal database directory if it doesn't exist.
This change adds logic to make sure the parent directory of the
internal database (that database dir in app's data) exists and is
created before proceeding to copy the pre-populated database.

Also re-enabled tests and added two not necessary but just in-case
delete database file test calls.

Bug: 134913247
Test: PrepackageTest in emulator with API 16, 17, 23, 24
Change-Id: I588bdfbf609708d05ba881496e6827a8170454cd