Move away from Ambients for parent data and position callbacks
Both parent data and position change callback notifications were
using ambients as part of their implementations. This CL changes
that to using emitted nodes, which should improve performance
of composition/recomposition.

Performance of composition has improved by over 35% for large
compositions.

It may be useful to use DataNode for things like TestTag, but
we would need to modify it a bit so that changes to TestTag
can be propagated to the SemanticsNode children.

Test: ParentDataTest, OnPositionedTest, ui-demo
Change-Id: I096f7129a2b87e74350927500a960e97896f3173