Allow single-value @Relation field.
This change lifts the restriction of @Relation annotated fields needing
to be List or Set allowing for a more seamless representation of a
one-to-one relationship.

Changes in the generated code is so that it is not always assumed that
the relation field is a collection, thus not needing a temporary
collection when collecting the relating objects.

Bug: 62905145
Test: ./gradlew room:room-compiler:test \
                room:integration-tests:room-testapp:cC

Change-Id: I462d59e89b55c43842df43cb84c1aeca0b2b743e